refactor(a11y): KZN-1111 Move package to components
