--- packages/a11y/index.ts
@@ -1,2 +1,0 @@
-export * from "./src/SkipLink"
-export * from "./src/VisuallyHidden/VisuallyHidden"
