--- packages/a11y/package.json
@@ -1,38 +1,0 @@
-{
-  "name": "@kaizen/a11y",
-  "version": "1.7.7",
-  "description": "A collection of accessibility utils",
-  "scripts": {
-    "prepublish": "tsc --project tsconfig.dist.json",
-    "build": "yarn clean && yarn prepublish",
-    "build:watch": "yarn clean && yarn prepublish --watch",
-    "clean": "rimraf '**/*.d.ts' '**/*.js' '**/*.map'"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/cultureamp/kaizen-design-system.git",
-    "directory": "packages/a11y"
-  },
-  "bugs": {
-    "url": "https://github.com/cultureamp/kaizen-design-system/issues"
-  },
-  "files": [
-    "**/*",
-    "!**/*.ts",
-    "!**/*.tsx",
-    "**/*.d.ts",
-    "!**/*.spec.*",
-    "!docs/**/*",
-    "!tsconfig.dist.json"
-  ],
-  "author": "",
-  "private": false,
-  "license": "MIT",
-  "peerDependencies": {
-    "react": "^16.14.0 || ^17.0.0 || ^18.0.0"
-  },
-  "dependencies": {
-    "@kaizen/component-base": "^1.1.5",
-    "@kaizen/component-library": "^16.5.3"
-  }
-}
