--- packages/a11y/tsconfig.dist.json
@@ -1,5 +1,0 @@
-{
-  "extends": "../../tsconfig.dist.json",
-  "include": ["index.ts", "src/**/*"],
-  "exclude": ["node_modules", "**/*.spec.ts", "**/*.spec.tsx"]
-}
