+++ packages/components/package.json
@@ -35,6 +35,7 @@
     "test": "jest"
   },
   "dependencies": {
+    "@kaizen/component-base": "^1.1.5",
     "@kaizen/component-library": "^16.5.3",
     "@kaizen/draft-divider": "^2.2.8",
     "@kaizen/draft-tooltip": "^5.4.37",
