+++ packages/components/src/a11y/index.ts
@@ -1,0 +1,2 @@
+export * from "./SkipLink"
+export * from "./VisuallyHidden/VisuallyHidden"
